requirements.txt
fastapi
uvicorn
gradio_client
pydantic
python-multipart
python-dotenv
slowapi
mtranslate
httpx
